test: extract MockArtifactsService.swift from ExportTestSupport.swift
Splits the primary `MockArtifactsService` test double (~45 lines) out
of ExportTestSupport.swift into its own file. ExportTestSupport.swift
retains the smaller Mock{ClipboardService, DebugBundleExporter,
PrivacyDataExporter} doubles.

Byte-identical move of lines 4-47. Both files retain their test
imports.

ExportTestSupport.swift: 98 → 52 lines. Tests: 667.
